Complete MinecraftProcessBuilder
